test: spell sparse-zerodata's NT ULONG fixed-width
`make ARCH=x86_64 asan` has been red on main with:

    FAIL zerodata: computed bytes-per-cluster is 0, which is not a
                   usable geometry

which reads like a filesystem with no geometry, or like src/ trusting a
zero it never validated. It was neither. fuzz/ntstubs.c:1888-1906 answers
FileFsSizeInformation with SectorsPerAllocationUnit=8 and
BytesPerSector=512 -- an ordinary 4096-byte cluster -- and both consumers
of those fields already validate (src/fcntl/fadvise.c:141-146 returns
LLONG_MAX on a zero cluster; src/stat/statvfs.c:95-120 only reports it,
after a range check, and never divides).

The bug was in this test's own hand-declared NT struct. It spelled the
two trailing fields of FILE_FS_SIZE_INFORMATION `unsigned long`, which is
32 bits under LLP64 -- correct for the tcc PE target, and 64 bits in the
native LP64 ELF build tools/asan-build.sh produces. That made the
structure 32 bytes rather than 24, so SectorsPerAllocationUnit read back
as the whole 64-bit word at offset 16 and BytesPerSector read past
everything the stub wrote.

The evidence is arithmetic rather than inference: the failing run printed
SectorsPerAllocationUnit=2199023255560, and (512 << 32) | 8 is exactly
2199023255560. The stub's own length check could not catch it -- the test
passed 32, which is >= sizeof *fs, so the call succeeded.

Spell it uint32_t, as src/internal/nt.h:36 does. A no-op for the PE
build, verified rather than assumed: rebuilt with x86_64-win32-tcc and
re-run under Wine, exit 77 and geometry 4096, unchanged. Same change also
fixes a second silent misread in this file, FILE_BASIC_INFORMATION's
FileAttributes, which was a 64-bit load straddling the end of the stub's
40-byte struct and now reads 0x20 (FILE_ATTRIBUTE_ARCHIVE).

Not masked: the g_cluster <= 0 guard is untouched and now passes on real
data, and the test goes on to put 16 rows through the actual FSCTL where
before it died on its first line. `make ARCH=x86_64 asan` goes from exit
1 to exit 0, 54/63 passed with sparse-zerodata moving FAIL -> UNVERIFIED.

The general hazard this is one instance of: test/*.c hand-declares the NT
surface and is compiled BOTH LLP64 (tcc) and LP64 (asan), so any NT
struct field spelled `unsigned long` is right in one and silently wrong
in the other, with no diagnostic from either compiler. src/internal/nt.h
uses fixed-width types and carries NT_LAYOUT_OFFSET assertions; test/ has
neither. test/posix-signal.c:71 has the same typedef -- harmless today
because it only passes scalars by value, and armed the moment an NT
struct joins that file.

+++ test/sparse-zerodata.c
@@ -140,6 +140,7 @@
 #include <errno.h>
 #include <fcntl.h>
 #include <unistd.h>
+#include <stdint.h>
 
 static int fails;
 static int unverified;
@@ -149,7 +150,17 @@ static int unverified;
 typedef int NTSTATUS;
 typedef void *HANDLE;
 typedef void *PVOID;
-typedef unsigned long ULONG;
+/* NT's ULONG is 32-bit on every target NT runs on, and `unsigned long`
+ * is 32-bit under LLP64 -- so spelling it that way is correct for the
+ * PE build and silently wrong anywhere `unsigned long` is 64 bits.  The
+ * native AddressSanitizer build (tools/asan-build.sh) is exactly that:
+ * LP64 ELF, where it made FILE_FS_SIZE_INFORMATION's two trailing
+ * ULONGs 8 bytes each, so SectorsPerAllocationUnit read back as
+ * (BytesPerSector << 32) | SectorsPerAllocationUnit and BytesPerSector
+ * read past the structure as 0 -- a bytes-per-cluster of 0 that looked
+ * like a filesystem with no geometry.  Fixed width, as
+ * src/internal/nt.h:36 spells it, is right on both. */
+typedef uint32_t ULONG;
 typedef unsigned char BOOLEAN;
 typedef long long LONGLONG;
 #ifdef __i386__
